Make HandlerScope internally constructed
